// test/integration/test/ports/_fixtures/issues-18598-select-count-true/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
type DefaultLiteralValue<CodecId extends string, _Encoded> = CodecId extends keyof CodecTypes
  ? CodecTypes[CodecId]['output']
  : _Encoded;

export type FieldOutputTypes = {
  readonly public: {
    readonly Post: { readonly id: Char<24>; readonly userId: CodecTypes['pg/text@1']['output'] };
    readonly User: { readonly id: Char<24> };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Post: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly User: { readonly id: CodecTypes['sql/char@1']['input'] };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly post: { readonly id: Char<24>; readonly userId: CodecTypes['pg/text@1']['output'] };
    readonly user: { readonly id: Char<24> };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly post: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly user: { readonly id: CodecTypes['sql/char@1']['input'] };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;